+++ src/lib/pages/proposal-details/components/index.ts
@@ -1,1 +1,3 @@
 export * from "./ProposalTop";
+export * from "./ProposalOverview";
+export * from "./VoteDetail";